Update vulkan bindings to 1.4.309.0 (#2)
* Update vulkan bindings to 1.4.309.0

* remove debug code

+++ QuantumBinding.Generator/Processors/CheckFlagsEnumPass.cs
@@ -20,7 +20,7 @@ namespace QuantumBinding.Generator.Processors
             {
                 try
                 {
-                    var value = Convert.ToInt32(item.Value);
+                    var value = Convert.ToInt64(item.Value);
 
                     if (value < 0)
                     {
